% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_partial_arc_lengths}
\alias{compute_partial_arc_lengths}
\title{Compute partial arc lengths}
\usage{
compute_partial_arc_lengths(coords)
}
\arguments{
\item{coords}{[nx2 matrix] Matrix of the points of the lines}
}
\description{
Computes the partial arc lengths from a matrix of coordinates of a SpatialLine
}
