\name{PCWprior-package}
\alias{PCWprior-package}
\alias{PCWprior}
\docType{package}
\title{
\packageTitle{PCWprior}
}
\description{
\packageDescription{PCWprior}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{PCWprior}
\packageIndices{PCWprior}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{PCWprior}

Maintainer: \packageMaintainer{PCWprior}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
