% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1D_analytic.R
<<<<<<<< Updated upstream:man/WCP_1D_GPtail_analytic.Rd
\name{WCP_1D_GPtail_analytic}
\alias{WCP_1D_GPtail_analytic}
|||||||| Stash base:man/WCP_1D_GP_analytic.Rd
\name{WCP_1D_GP_analytic}
\alias{WCP_1D_GP_analytic}
========
\name{WCP1_1D_GPtail_analytic}
\alias{WCP1_1D_GPtail_analytic}
>>>>>>>> Stashed changes:man/WCP1_1D_GPtail_analytic.Rd
\title{The 1d analytic density WCP prior for xi of generalized Pareto distribution}
\usage{
<<<<<<<< Updated upstream:man/WCP_1D_GPtail_analytic.Rd
WCP_1D_GPtail_analytic(seq_xi, eta, inla_table = FALSE)
|||||||| Stash base:man/WCP_1D_GP_analytic.Rd
WCP_1D_GP_analytic(seq_xi, eta)
========
WCP1_1D_GPtail_analytic(seq_xi, eta, inla_table = FALSE)
>>>>>>>> Stashed changes:man/WCP1_1D_GPtail_analytic.Rd
}
\arguments{
\item{seq_xi}{A vector of values of xi parameter.}

\item{eta}{User specified parameter of the WCP prior.}

\item{inla_table}{Should the results be returned as a table to be used in INLA? Default is FALSE.}
}
\value{
A list including densities on seq_xi and a inla table
}
\description{
The 1d analytic density WCP prior for xi of generalized Pareto distribution
}
