% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2D_analytic.R
\name{WCP1_2D_GP_analytic}
\alias{WCP1_2D_GP_analytic}
\title{The 2d analytic density WCP prior for sigma and xi of generalized Pareto distribution}
\usage{
WCP1_2D_GP_analytic(sigma, xi, eta)
}
\arguments{
\item{sigma}{Sigma parameter.}

\item{xi}{Xi parameter.}

\item{eta}{User specified parameter of the WCP prior.}
}
\value{
A value of density evaluated at c(sigma, xi).
}
\description{
The 2d analytic density WCP prior for sigma and xi of generalized Pareto distribution
}
