% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1D_analytic.R
\name{WCP2_1D_Gaussian_mean_analytic}
\alias{WCP2_1D_Gaussian_mean_analytic}
\title{The 1d analytic density WCP2 prior for mean of Gaussian distribution}
\usage{
WCP2_1D_Gaussian_mean_analytic(seq_m, eta, inla_table = FALSE)
}
\arguments{
\item{seq_m}{A vector of values of mean parameter.}

\item{eta}{User specified parameter of the WCP prior.}

\item{inla_table}{Should the results be returned as a table to be used in INLA? Default is FALSE.}
}
\value{
A list including densities on seq_m and a inla table
}
\description{
The 1d analytic density WCP2 prior for mean of Gaussian distribution
}
