% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1D_analytic.R
\name{WCP_1D_AR1_analytic}
\alias{WCP_1D_AR1_analytic}
\title{The 1d analytic density WCP prior for phi of stationary AR1 process}
\usage{
WCP_1D_AR1_analytic(phi, eta, n, sigma)
}
\arguments{
\item{phi}{phi parameter.}

\item{eta}{User specified parameter of the WCP prior.}

\item{n}{Length of AR1 process.}

\item{sigma}{Standard deviation of the process.}
}
\value{
A value of density evaluated at c(mean, std).
}
\description{
The 1d analytic density WCP prior for phi of stationary AR1 process
}
