% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1D_analytic.R
\name{WCP_1D_GP_analytic}
\alias{WCP_1D_GP_analytic}
\title{The 1d analytic density WCP prior for xi of generalized Pareto distribution}
\usage{
WCP_1D_GP_analytic(seq_xi, eta)
}
\arguments{
\item{seq_xi}{A vector of values of xi parameter.}

\item{eta}{User specified parameter of the WCP prior.}
}
\value{
A list including densities on seq_xi and a inla table
}
\description{
The 1d analytic density WCP prior for xi of generalized Pareto distribution
}
