% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1D_numerical.R
\name{WCP_1D_Numerical_Density}
\alias{WCP_1D_Numerical_Density}
\title{The 1d numerical density WCP prior}
\usage{
WCP_1D_Numerical_Density(
  base_theta,
  L,
  L_included = TRUE,
  U,
  U_included = FALSE,
  W_func,
  eta,
  cutoff1,
  cutoff2 = NULL,
  mesh_width
)
}
\arguments{
\item{base_theta}{Base model theta.}

\item{L}{Lower bound of theta.}

\item{L_included}{If L should be included in the domain, a TRUE or FALSE value}

\item{U}{Upper bound of theta.}

\item{U_included}{If U should be included in the domain, a TRUE or FALSE value}

\item{W_func}{A function that returns the Wasserstein distance given theta.}

\item{eta}{User specified parameter of the WCP prior.}

\item{cutoff1}{Cutoff parameter that decide upper bound of the Wasserstein distance on the right hand side of base_theta.}

\item{cutoff2}{Cutoff parameter that decide upper bound of the Wasserstein distance on the left hand side of base_theta.}

\item{mesh_width}{Mesh width.}
}
\value{
A list of density locations and densities evaluated on those locations.
}
\description{
The 1d numerical density WCP prior
}
