% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2D_numerical.R
\name{WCP_2D_Numerical_Density}
\alias{WCP_2D_Numerical_Density}
\title{The 2d numerical approximation of WCP prior}
\usage{
WCP_2D_Numerical_Density(
  W_func,
  eta,
  mesh_width = NULL,
  alpha = 1,
  tau = mesh_width/1000,
  cutoff = 0.01,
  region = list(type = "conic", lower_angle = 0, upper_angle = pi, base_theta = c(0, 0)),
  lc_multiplier = 20,
<<<<<<< Updated upstream
  parallel = FALSE,
  NumCores = parallel::detectCores() - 1
||||||| Stash base
  NumCores
=======
  parallel = FALSE,
  NumCores = parallel::detectCores() - 1,
  visual_mesh = FALSE
>>>>>>> Stashed changes
)
}
\arguments{
\item{W_func}{A function that returns the Wasserstein distance given theta1 and theta2.}

\item{eta}{User specified parameter of the WCP prior.}

\item{mesh_width}{Mesh width.}

\item{alpha}{Power of mesh width for a finer mesh.}

\item{tau}{tau parameter, indicating how much bounded away from boundary.}

\item{cutoff}{Cutoff parameter indicating probability tail mass for the Wasserstein distance.}

\item{region}{A list that specify region information.}

\item{lc_multiplier}{Multiplier determines number of level curves}

\item{parallel}{A logic value that indicating whether the user wants to run the function with multiple cpu.}

\item{NumCores}{Number of cores to run the function.}

\item{visual_mesh}{A logic value to determin whether to plot the mesh.}
}
\value{
A list of density locations, densities evaluated on those locations and other utilities that can help access accuracy.
}
\description{
The 2d numerical approximation of WCP prior
}
