% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1D_analytic.R
\name{WCP1_1D_GPtail_analytic}
\alias{WCP1_1D_GPtail_analytic}
\title{The 1d analytic density WCP prior for xi of generalized Pareto distribution}
\usage{
WCP1_1D_GPtail_analytic(seq_xi, eta, inla_table = FALSE)
}
\arguments{
\item{seq_xi}{A vector of values of xi parameter.}

\item{eta}{User specified parameter of the WCP prior.}

\item{inla_table}{Should the results be returned as a table to be used in INLA? Default is FALSE.}
}
\value{
A list including densities on seq_xi and a inla table
}
\description{
The 1d analytic density WCP prior for xi of generalized Pareto distribution
}
