% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1D_analytic.R
\name{WCP_1D_Gaussian_precision_analytic}
\alias{WCP_1D_Gaussian_precision_analytic}
\title{The 1d analytic density WCP2 prior for tau (1/variance) of Gaussian distribution}
\usage{
WCP_1D_Gaussian_precision_analytic(seq_tau, eta, inla_table = FALSE)
}
\arguments{
\item{seq_tau}{A vector of values of tau parameter.}

\item{eta}{User specified parameter of the WCP prior.}

\item{inla_table}{Should the results be returned as a table to be used in INLA? Default is FALSE.}
}
\value{
A list including densities on seq_tau and a inla table
}
\description{
The 1d analytic density WCP2 prior for tau (1/variance) of Gaussian distribution
}
