% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1D_analytic.R
\name{WCP_1D_GP_analytic}
\alias{WCP_1D_GP_analytic}
\title{The 1d analytic density WCP prior for xi of generalized Pareto distribution}
\usage{
WCP_1D_GP_analytic(xi, eta)
}
\arguments{
\item{xi}{Xi parameter.}

\item{eta}{User specified parameter of the WCP prior.}
}
\value{
A value of density evaluated at c(sigma, xi).
}
\description{
The 1d analytic density WCP prior for xi of generalized Pareto distribution
}
