% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1D_numerical.R
\name{WCP_1D_Numerical_Density}
\alias{WCP_1D_Numerical_Density}
\title{The 1d numerical density WCP prior}
\usage{
WCP_1D_Numerical_Density(
  W_func,
  eta,
  base_theta,
  L,
  U,
  mesh_width = NULL,
  mesh_n = 100,
  cutoff1 = 0.01,
  cutoff2 = NULL,
  L_included = FALSE,
  U_included = FALSE,
  inla_table = FALSE
)
}
\arguments{
\item{W_func}{A function that returns the Wasserstein distance given theta.}

\item{eta}{User specified parameter of the WCP prior.}

\item{base_theta}{Base model theta.}

\item{L}{Lower bound of theta.}

\item{U}{Upper bound of theta.}

\item{mesh_width}{Mesh width to be used. If not used, mesh_n will be used.}

\item{mesh_n}{Number of mesh nodes to be used. Will not be used if mesh_width is given.}

\item{cutoff1}{Cutoff parameter that decide upper bound of the Wasserstein distance on the right hand side of base_theta.}

\item{cutoff2}{Cutoff parameter that decide upper bound of the Wasserstein distance on the left hand side of base_theta.}

\item{L_included}{If L should be included in the domain, a TRUE or FALSE value}

\item{U_included}{If U should be included in the domain, a TRUE or FALSE value}

\item{inla_table}{Should the results be returned as a table to be used in INLA? Default is FALSE.}
}
\value{
A list of density locations and densities evaluated on those locations.
}
\description{
The 1d numerical density WCP prior
}
