% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2D_numerical.R
\name{WCP_2D_Numerical_Density}
\alias{WCP_2D_Numerical_Density}
\title{The 2d numerical approximation of WCP prior}
\usage{
WCP_2D_Numerical_Density(
  boundary_path,
  W_func,
  mesh_width,
  alpha = 1,
  W_lower_bound,
  W_upper_bound,
  eta,
  L2,
  U2,
  level_curve_type,
  lc_multiplier = 20,
  NumCores
)
}
\arguments{
\item{boundary_path}{A sequence of points that represent boundary of computational domain.}

\item{W_func}{A function that returns the Wasserstein distance given theta1 and theta2.}

\item{mesh_width}{Mesh width.}

\item{alpha}{Power of mesh width for a finer mesh.}

\item{W_lower_bound}{The lower bound of the Wasserstein distance.}

\item{W_upper_bound}{The upper bound of the Wasserstein distance.}

\item{eta}{User specified parameter of the WCP prior.}

\item{L2}{Lower bound of theta 2.}

\item{U2}{Upper bound of theta 2.}

\item{level_curve_type}{Type of level curve, the value should be either LL or LU}

\item{lc_multiplier}{Multiplier determines number of level curves}

\item{NumCores}{Number of cores to run the function.}
}
\value{
A list of density locations, densities evaluated on those locations and other utilities that can help access accuracy.
}
\description{
The 2d numerical approximation of WCP prior
}
