% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2D_numerical.R
\name{WCP_2D_Numerical_Density}
\alias{WCP_2D_Numerical_Density}
\title{The 2d numerical approximation of WCP prior}
\usage{
WCP_2D_Numerical_Density(
  W_func,
  eta,
  mesh_width,
  alpha = 1,
  tau = mesh_width/1000,
  cutoff = 0.01,
  region = list(type = "conic", lower_angle = 0, upper_angle = pi, base_theta = c(0, 0)),
  lc_multiplier = 20,
  buff_boundary = FALSE,
  parallel = FALSE,
  NumCores = parallel::detectCores() - 1,
  visual_mesh = FALSE
)
}
\arguments{
\item{W_func}{A function that returns the Wasserstein distance given theta1 and theta2.}

\item{eta}{User specified parameter of the WCP prior.}

\item{mesh_width}{Mesh width.}

\item{alpha}{Power of mesh width for a finer mesh.}

\item{tau}{tau parameter, indicating how much bounded away from boundary.}

\item{cutoff}{Cutoff parameter indicating probability tail mass for the Wasserstein distance.}

\item{region}{A list that specify region information. For example, for strip regions (regions of the type IxJ, where I = (a,b) and J = (c,d) are intervals, with one of them being unbounded), one can set `region = list(type = "strip", base_theta = c(a,c), corners = c(a,b,c,d))`, where some of the values `a`,`b`,`c` or `d` can be `-Inf` or `Inf`. If the type is conic, the list must be set, for example, as list(type = 'conic', lower_angle = 0, upper_angle = pi, base_theta = c(0,0)), where `lower_angle` and `upper_angle` contains the smaller and larger angles that determine the cone.}

\item{lc_multiplier}{Multiplier determines number of level curves}

\item{buff_boundary}{Should an additional buffer (of the magnitude of mesh_width) be added to the region to bound it away from the boundary of the domain?}

\item{parallel}{A logic value that indicating whether the user wants to run the function with multiple cpu.}

\item{NumCores}{Number of cores to run the function.}

\item{visual_mesh}{A logic value to determin whether to plot the mesh.}
}
\value{
A list of density locations, densities evaluated on those locations and other utilities that can help access accuracy.
}
\description{
The 2d numerical approximation of WCP prior
}
